java

+++ src/Malik/Calculatingclass.java
@@ -9,7 +9,7 @@ public class Calculatingclass {
         int number = input.nextInt();
         int count = 1;
         int sum = 0;
-        int average;
+        double average;
     while (count <= 10){
         count ++;
         sum += number;
@@ -18,9 +18,9 @@ public class Calculatingclass {
             number = input.nextInt();
         }
         }
-        average = sum / count;
+        average = (double) sum / count;
         System.out.printf("The sum of grade is %d", sum);
-        System.out.printf("\nThe average grade is %d", average);
+        System.out.printf("\nThe average grade is %f", average);
 
     }
 }
